Support emulated aarch64 builds
A new class, Builder_emuarm64 handles this in prepare_cloud_image.
Now .kitchen.yml is created when preparing the image, and there is
only one VM listed, corresponding to the prepared image.

With only one VM buildbot doesn't need to give a vm name to
kitchen. However instance() needs to return something for the
yield and error handling to work. Also, using the vm name avoids
a bunch of changes to lines running kitchen.